move Core/IdeaColl/Workbench to Core/Workbench
