+++ src/adhocracy_frontend/adhocracy_frontend/static/js/Packages/Core/IdeaCollection/Module.ts
@@ -1,7 +1,6 @@
 import * as AdhIdeaCollectionPollModule from "./Poll/Module";
 import * as AdhIdeaCollectionProcessModule from "./Process/Module";
 import * as AdhIdeaCollectionProposalModule from "./Proposal/Module";
-import * as AdhIdeaCollectionWorkbenchModule from "./Workbench/Module";
 
 export var moduleName = "adhIdeaCollection";
 
@@ -9,13 +8,11 @@ export var register = (angular) => {
     AdhIdeaCollectionPollModule.register(angular);
     AdhIdeaCollectionProcessModule.register(angular);
     AdhIdeaCollectionProposalModule.register(angular);
-    AdhIdeaCollectionWorkbenchModule.register(angular);
 
     angular
         .module(moduleName, [
             AdhIdeaCollectionPollModule.moduleName,
             AdhIdeaCollectionProcessModule.moduleName,
-            AdhIdeaCollectionProposalModule.moduleName,
-            AdhIdeaCollectionWorkbenchModule.moduleName
+            AdhIdeaCollectionProposalModule.moduleName
         ]);
 };
