+++ src/adhocracy_frontend/adhocracy_frontend/static/js/Packages/Core/Workbench/AddProposalButton.html
