+++ src/meinberlin_lib/meinberlin_lib/static/js/Packages/Meinberlin/CollaborativeText/Module.ts
@@ -5,10 +5,10 @@ import * as AdhProcessModule from "../../Core/Process/Module";
 import * as AdhResourceAreaModule from "../../Core/ResourceArea/Module";
 
 import * as AdhDocument from "../../Core/Document/Document";
-import * as AdhIdeaCollectionWorkbench from "../../Core/IdeaCollection/Workbench/Workbench";
 import * as AdhNames from "../../Core/Names/Names";
 import * as AdhProcess from "../../Core/Process/Process";
 import * as AdhResourceArea from "../../Core/ResourceArea/ResourceArea";
+import * as AdhWorkbench from "../../Core/Workbench/Workbench";
 
 import RIDocument from "../../../Resources_/adhocracy_core/resources/document/IDocument";
 import RIDocumentVersion from "../../../Resources_/adhocracy_core/resources/document/IDocumentVersion";
@@ -28,12 +28,12 @@ export var register = (angular) => {
             AdhResourceAreaModule.moduleName,
         ])
         .config(["adhResourceAreaProvider", "adhConfig", (adhResourceAreaProvider: AdhResourceArea.Provider, adhConfig) => {
-            AdhIdeaCollectionWorkbench.registerCommonRoutesFactory(
+            AdhWorkbench.registerCommonRoutesFactory(
                 RICollaborativeTextProcess, RIDocument, RIDocumentVersion)()(adhResourceAreaProvider);
-            AdhIdeaCollectionWorkbench.registerDocumentRoutesFactory(
+            AdhWorkbench.registerDocumentRoutesFactory(
                 RICollaborativeTextProcess, RIDocument, RIDocumentVersion)()(adhResourceAreaProvider);
 
-            var processHeaderSlot = adhConfig.pkg_path + AdhIdeaCollectionWorkbench.pkgLocation + "/AddDocumentSlot.html";
+            var processHeaderSlot = adhConfig.pkg_path + AdhWorkbench.pkgLocation + "/AddDocumentSlot.html";
             adhResourceAreaProvider.processHeaderSlots[processType] = processHeaderSlot;
         }])
         .config(["adhConfig", "adhProcessProvider", (adhConfig, adhProcessProvider : AdhProcess.Provider) => {
