+++ src/meinberlin_lib/meinberlin_lib/static/js/Packages/Meinberlin/IdeaCollection/Module.ts
@@ -4,10 +4,10 @@ import * as AdhProcessModule from "../../Core/Process/Module";
 import * as AdhResourceAreaModule from "../../Core/ResourceArea/Module";
 
 import * as AdhIdeaCollectionProposal from "../../Core/IdeaCollection/Proposal/Proposal";
-import * as AdhIdeaCollectionWorkbench from "../../Core/IdeaCollection/Workbench/Workbench";
 import * as AdhNames from "../../Core/Names/Names";
 import * as AdhProcess from "../../Core/Process/Process";
 import * as AdhResourceArea from "../../Core/ResourceArea/ResourceArea";
+import * as AdhWorkbench from "../../Core/Workbench/Workbench";
 
 import RIGeoProposal from "../../../Resources_/adhocracy_core/resources/proposal/IGeoProposal";
 import RIGeoProposalVersion from "../../../Resources_/adhocracy_core/resources/proposal/IGeoProposalVersion";
@@ -27,12 +27,12 @@ export var register = (angular) => {
             AdhResourceAreaModule.moduleName,
         ])
         .config(["adhResourceAreaProvider", "adhConfig", (adhResourceAreaProvider: AdhResourceArea.Provider, adhConfig) => {
-            AdhIdeaCollectionWorkbench.registerCommonRoutesFactory(
+            AdhWorkbench.registerCommonRoutesFactory(
                 RIIdeaCollectionProcess, RIGeoProposal, RIGeoProposalVersion)()(adhResourceAreaProvider);
-            AdhIdeaCollectionWorkbench.registerProposalRoutesFactory(
+            AdhWorkbench.registerProposalRoutesFactory(
                 RIIdeaCollectionProcess, RIGeoProposal, RIGeoProposalVersion, true)()(adhResourceAreaProvider);
 
-            var processHeaderSlot = adhConfig.pkg_path + AdhIdeaCollectionWorkbench.pkgLocation + "/ProcessHeaderSlot.html";
+            var processHeaderSlot = adhConfig.pkg_path + AdhWorkbench.pkgLocation + "/ProcessHeaderSlot.html";
             adhResourceAreaProvider.processHeaderSlots[processType] = processHeaderSlot;
         }])
         .config(["adhConfig", "adhProcessProvider", (adhConfig, adhProcessProvider : AdhProcess.Provider) => {
